Resolver testing is proving to be a bit ineffective. Its a bit too tightly coupled with depsolver but cant change it at the moment
